Fix macOS build and resolve a batch of Tier B/C warnings
- src/qcommon/net_ip.c: Define _DARWIN_C_SOURCE on macOS, move sys/types.h up, and add fallbacks for IFF_UP and struct ipv6_mreq to fix build breaks.
- src/botlib/l_script.c: Add printf format attributes to ScriptWarning and ScriptError to fix -Wformat-nonliteral.
- src/renderers/openglrenderer/tr_animation.c: Fix -Wcast-qual by adding const to unsigned short pointer casts.
- src/client/mp3: Add (void) to non-prototype declarations and mark unused dummy functions with __attribute__((unused)).
- src/client: Resolve multiple -Wsign-compare warnings in cl_parse.c, cl_ui.c, and cl_input.c by adding appropriate casts.

+++ src/client/cl_parse.c
@@ -260,7 +260,7 @@ static void CL_ParseSnapshot( msg_t *msg ) {
 	// read areamask
 	newSnap.areabytes = MSG_ReadByte( msg );
 
-	if ( newSnap.areabytes > sizeof(newSnap.areamask) )
+	if ( newSnap.areabytes > (int)sizeof(newSnap.areamask) )
 	{
 		Com_Error( ERR_DROP,"CL_ParseSnapshot: Invalid size %d for areamask", newSnap.areabytes );
 		return;
@@ -432,7 +432,7 @@ void CL_SystemInfoChanged( qboolean onlyGame ) {
 			continue; // already processed
 		}
 
-		if ( ( cvar_flags = Cvar_Flags( key ) ) == CVAR_NONEXISTENT )
+		if ( ( cvar_flags = (int)Cvar_Flags( key ) ) == (int)CVAR_NONEXISTENT )
 			Cvar_Get( key, value, CVAR_SERVER_CREATED | CVAR_ROM );
 		else
 		{
